Holy smokes got that can of beans organized, now to clean up.
