+++ Map_Project/tests/test_game.py
@@ -2,8 +2,9 @@ import unittest
 from unittest.mock import patch
 
 from classes.game import Game
-from classes.plain import Plain
-from classes.maptile import MapTile
+# from classes.plain import Plain
+# from classes.maptile import MapTile
+from classes.mapengine import MapEngine
 from classes.plebian import Plebian
 from classes.jester import Jester
 
@@ -16,7 +17,12 @@ class TestGame(unittest.TestCase):
           Reference this:
           https://stackoverflow.com/questions/47690020/python-3-unit-tests-with-user-input
     '''
-    @patch('classes.game.input', create=True)
+    """
+    run tests. "run" as in, the "run" method for the Game Class.
+    """
+    # @patch('classes.game.input', create=True) # mocked_input_mapengine, mocked_input_game
+    # @patch('classes.mapengine.input', create=True)
+    @patch('builtins.input', create=True)
     def test_run_success_case(self, mocked_input):
         """
         Go through all of the different commands that we have for the "run" method and see that
@@ -36,6 +42,13 @@ class TestGame(unittest.TestCase):
         move = "move"
         up = "up"
         terminate = "terminate"
+        # mocked_input_mapengine.side_effect = [
+        #     map_width,
+        #     map_height,
+        #     # Create Character Command
+        #     x,
+        #     y,
+        # ]
         mocked_input.side_effect = [
             map_width,
             map_height,
@@ -61,13 +74,15 @@ class TestGame(unittest.TestCase):
         crazy_game.run()
 
         # After all of that tom-foolery we should have a map attached to the game.
-        self.assertNotEqual(crazy_game.map, None)
+        self.assertNotEqual(crazy_game.map_engine.map, None)
         self.assertNotEqual(
-            crazy_game.map.grid[int(x)][int(y) + 1].occupant,
+            crazy_game.map_engine.map.grid[int(x)][int(y) + 1].occupant,
             None
         )
 
-    @patch('classes.game.input', create=True)
+    # @patch('classes.game.input', create=True) # mocked_input_mapengine, mocked_input_game
+    # @patch('classes.mapengine.input', create=True)
+    @patch('builtins.input', create=True)
     def test_run_unrecognized_input_case(self, mocked_input):
         """
         The test case where we try to feed the interpreter some unrecognized inputs.
@@ -78,6 +93,10 @@ class TestGame(unittest.TestCase):
         map_height = "10"
         unrecognized_input = "And the flames will rise!"
         terminate = "terminate"
+        # mocked_input_mapengine.side_effect = [
+        #     map_width,
+        #     map_height,
+        # ]
         mocked_input.side_effect = [
             map_width,
             map_height,
@@ -90,357 +109,357 @@ class TestGame(unittest.TestCase):
 
         # The game should run (and close), and with this pass we should only have the
         # map set for game, so it should NOT be "None".
-        self.assertNotEqual(super_game.map, None)
-
-    """
-    set map tests.
-    """
-    def test_set_map_success_case(self):
-        chump_game = Game()
-        # New Game should have no map set yet.
-        self.assertEqual(chump_game.map, None)
-
-        # Note that a Plain IS a map. Specifically, a
-        # Plain is a subclass of Map.
-        some_plain = Plain()
-        chump_game.set_map(some_plain)
-        self.assertEqual(chump_game.map, some_plain)
-
-    def test_set_map_failure_case(self):
-        chump_game = Game()
-        # New Game should have no map set yet.
-        self.assertEqual(chump_game.map, None)
-
-        # Unfortunately we can't attach sexy beasts to games.
-        with self.assertRaises(TypeError) as context:
-            chump_game.set_map("Hi I'm a sexy beast.")
-
-            self.assertTrue("Unable to attach Map Type obect to Game." in context.exception)
-
-    """
-    add_entity_to_map tests.
-    """
-    def test_add_entity_to_map_no_map_case(self):
-        '''
-        This should fail as we can't add an entity to a map
-        that doesn't exist.
-        '''
-        chump_game = Game()
-        # Important to note that anyone can be a sexy beast, even the plebians.
-        sexy_beast = Plebian()
-
-        with self.assertRaises(ValueError) as context:
-            chump_game.add_entity_to_map(sexy_beast)
-
-            self.assertTrue("There is no map associated with the game!" in context.exception)
-
-    def test_add_entity_to_map_non_entity_case(self):
-        '''
-        This should fail as we can't add something to the map that does not
-        have an x and y cordinate.
-        '''
-        chump_game = Game()
-        some_plain = Plain()
-        some_plain.initialize(10, 10)
-        chump_game.set_map(some_plain)
-
-        with self.assertRaises(AttributeError) as context:
-            chump_game.add_entity_to_map(42 + 300)
-
-            self.assertTrue("'int' object has no attribute 'x'" in context.exception)
-
-    def test_add_entity_to_map_SUCCESS_case(self):
-        '''
-        Default success case. We should be able to add a character to a map, no problems.
-        '''
-        chump_game = Game()
-        some_plain = Plain()
-        some_plain.initialize(10, 10)
-        chump_game.set_map(some_plain)
-
-        x_cord = 2
-        y_cord = 2
-
-        the_victim = Plebian()
-        chump_game.set_entity_position(the_victim, x_cord, y_cord)
-        chump_game.add_entity_to_map(the_victim)
-
-        self.assertEqual(
-            chump_game.map.grid[x_cord][y_cord].occupant,
-            the_victim
-        )
-
-    def test_add_entity_to_map_DUPLICATE_case(self):
-        '''
-        We're going to add the same character TWICE to the same map.
-        Once with the same coordinates, and then once with different coordinates.
-
-        The first case should fail (due to two entities being on the same space),
-        but the second case should pass (due to them being on different spaces).
-
-        I'm okay (for now) with the second case passing because maybe there could
-        be "clones" on the map, but this is worth investigation. And I dunno, it
-        could get messy.
-        '''
-        chump_game = Game()
-        some_plain = Plain()
-        some_plain.initialize(10, 10)
-        chump_game.set_map(some_plain)
-
-        x_cord = 2
-        y_cord = 2
-
-        the_victim = Plebian()
-        chump_game.set_entity_position(the_victim, x_cord, y_cord)
-        chump_game.add_entity_to_map(the_victim)
-
-        self.assertEqual(
-            chump_game.map.grid[x_cord][y_cord].occupant,
-            the_victim
-        )
-
-        with self.assertRaises(AttributeError) as context:
-            chump_game.add_entity_to_map(the_victim)
-
-            self.assertTrue("Map space is already occupied!" in context.exception)
-
-        # Update the position so we can re-add "the victim" to the map.
-        chump_game.set_entity_position(the_victim, x_cord + 1, y_cord + 1)
-        chump_game.add_entity_to_map(the_victim)
-
-        self.assertEqual(
-            chump_game.map.grid[x_cord + 1][y_cord + 1].occupant,
-            the_victim
-        )
-        # Check again to make sure the first "victim" is still on the map.
-        self.assertEqual(
-            chump_game.map.grid[x_cord][y_cord].occupant,
-            the_victim
-        )
-
-    def test_add_entity_to_map_OUT_OF_BOUNDS_case(self):
-        '''
-        Try to add an entity to a map, but the entity coordinates will
-        be out of bounds.
-
-        We'll get a key error, but maybe there'd be a way to have more
-        rigorous error checking. Maybe have some checks on the Character class?
-        Not sure.
-        '''
-        chump_game = Game()
-        some_plain = Plain()
-        some_plain.initialize(10, 10)
-        chump_game.set_map(some_plain)
-
-        x_cord = 12
-        y_cord = 12
-
-        the_lost_one = Plebian()
-        chump_game.set_entity_position(the_lost_one, x_cord, y_cord)
-        with self.assertRaises(KeyError):
-            chump_game.add_entity_to_map(the_lost_one)
-
-    """
-    remove_entity_from_map tests.
-    """
-    def test_remove_entity_from_map_SUCCESS_case(self):
-        '''
-        Put an entity on a map and then remove it. This should be pretty
-        straightforward.
-        '''
-        chump_game = Game()
-        some_plain = Plain()
-        some_plain.initialize(10, 10)
-        chump_game.set_map(some_plain)
-
-        x_cord = 2
-        y_cord = 2
-
-        the_victim = Plebian()
-        chump_game.set_entity_position(the_victim, x_cord, y_cord)
-        chump_game.add_entity_to_map(the_victim)
-
-        # Now we remove the victim. I feel like I'm some mafia member cleaning the scene.
-        chump_game.remove_entity_from_map(the_victim)
-        self.assertEqual(
-            chump_game.map.grid[x_cord][y_cord].occupant,
-            None
-        )
-
-    @patch('builtins.print')
-    def test_remove_entity_from_map_EMPTY_case(self, mock_print):
-        '''
-        Create an entity, but don't add it to the map. We'll try to remove it, but won't
-        be able to as, well, it just doesn't exist.
-        '''
-        chump_game = Game()
-        some_plain = Plain()
-        some_plain.initialize(10, 10)
-        chump_game.set_map(some_plain)
-
-        x_cord = 2
-        y_cord = 2
-
-        mr_waffles = Plebian()
-        chump_game.set_entity_position(mr_waffles, x_cord, y_cord)
-
-        chump_game.remove_entity_from_map(mr_waffles)
-        mock_print.assert_called_with("{}, {} is already empty!".format(x_cord, y_cord))
-
-    def test_remove_entity_from_map_NO_MAP_case(self):
-        '''
-        This should fail, as how can we remove something from a map if a map doesn't exist?
-
-        At this point though I feel like I'm testing the compiler more than anything else.
-        '''
-        chump_game = Game()
-
-        x_cord = 2
-        y_cord = 2
-
-        some_guy = Plebian()
-        chump_game.set_entity_position(some_guy, x_cord, y_cord)
-
-        with self.assertRaises(ValueError) as context:
-            chump_game.remove_entity_from_map(some_guy)
-
-            self.assertTrue("There is no map associated with the game!" in context.exception)
-
-    def test_remove_entity_from_map_non_entity_case(self):
-        '''
-        This will obviously error. I can't think of many cases where this should naturally
-        occur though.
-        '''
-        chump_game = Game()
-        some_plain = Plain()
-        some_plain.initialize(10, 10)
-        chump_game.set_map(some_plain)
-
-        with self.assertRaises(AttributeError) as context:
-            chump_game.remove_entity_from_map(666)
-
-            self.assertTrue("'int' object has no attribute 'x'" in context.exception)
-
-    """
-    check_if_occupied tests.
-    """
-    def test_check_if_occupied_NO_MAP_case(self):
-        '''
-        Hard to check if a tile on a map is occupied if there's no map.
-
-        I'm sure someone has said this before. Maybe the voices in my head.
-        '''
-        chump_game = Game()
-        x_cord = 2
-        y_cord = 2
-        with self.assertRaises(ValueError) as context:
-            chump_game.check_if_occupied(x_cord, y_cord)
-
-            self.assertTrue("There is no map associated with the game!" in context.exception)
-
-    def test_check_if_occupied_OUT_OF_BOUNDS_case(self):
-        """
-        Exactly what is described by the title. We should get a key error.
-        """
-        the_game = Game()
-        smallest_plain = Plain()
-        smallest_plain.initialize(3, 3)
-        the_game.set_map(smallest_plain)
-
-        x_cord = 40000
-        y_cord = 40000
-
-        with self.assertRaises(KeyError):
-            the_game.check_if_occupied(x_cord, y_cord)
-
-    def test_check_if_occupied_FALSE_case(self):
-        """
-        Check if an unoccupied square is unoccupied.
-
-        Remember, "False" isn't necessarily a bad thing.
-
-        "Did you murder this man?"
-
-        "False."
-        """
-        chump_game = Game()
-        plane = Plain()
-        plane.initialize(10, 10)
-        chump_game.set_map(plane)
-
-        x_cord = 8
-        y_cord = 8
-
-        self.assertEqual(
-            chump_game.check_if_occupied(x_cord, y_cord),
-            False
-        )
-
-    def test_check_if_occupied_TRUE_case(self):
-        """
-        Check if an occupied square is occupied.
-
-        Note that "True" CAN be a bad thing.
-
-        "Did you set your boss's house on fire?"
-
-        "True."
-        """
-        chump_game = Game()
-        the_void = Plain()
-        the_void.initialize(10, 10)
-        chump_game.set_map(the_void)
-
-        x_cord = 3
-        y_cord = 3
-
-        master_of_the_universe = Plebian()
-        chump_game.set_entity_position(master_of_the_universe, x_cord, y_cord)
-        chump_game.add_entity_to_map(master_of_the_universe)
-
-        self.assertEqual(
-            chump_game.check_if_occupied(x_cord, y_cord),
-            True
-        )
-
-    """
-    initialize_map tests.
-    """
-    @patch('classes.game.input', create=True)
-    def test_initialize_map_SUCCESS_case(self, mocked_input):
-        """
-        Simple case of initializing a map and assigning it to the game.
-
-        Note we're going to have to supply user inputs for this.
-
-        Furthermore, this is just to make sure the map gets attached to the game,
-        we can be more rigorous with the "initialize" method on the Maps.
-        """
-        x_cord = 10
-        y_cord = 8
-        mocked_input.side_effect = [x_cord, y_cord]
-        a_game = Game()
-        a_game.initialize_map()
-
-        self.assertTrue(isinstance(a_game.map, Plain))
-        self.assertEqual(a_game.map.width, x_cord)
-        self.assertEqual(a_game.map.height, y_cord)
-
-    @patch('classes.game.input', create=True)
-    def test_initialize_map_FAILURE_case(self, mocked_input):
-        """
-        Going to crash this method by giving it some string inputs.
-
-        This raises the question of maybe we should allow the user to try
-        again if they input something "dumb", but that might be a later task.
-        """
-        x_cord = "ten"
-        y_cord = "eight"
-        mocked_input.side_effect = [x_cord, y_cord]
-        a_game = Game()
-
-        with self.assertRaises(ValueError):
-            a_game.initialize_map()
+        self.assertNotEqual(super_game.map_engine.map, None)
+
+    # """
+    # set map tests.
+    # """
+    # def test_set_map_success_case(self):
+    #     chump_game = Game()
+    #     # New Game should have no map set yet.
+    #     self.assertEqual(chump_game.map, None)
+
+    #     # Note that a Plain IS a map. Specifically, a
+    #     # Plain is a subclass of Map.
+    #     some_plain = Plain()
+    #     chump_game.set_map(some_plain)
+    #     self.assertEqual(chump_game.map, some_plain)
+
+    # def test_set_map_failure_case(self):
+    #     chump_game = Game()
+    #     # New Game should have no map set yet.
+    #     self.assertEqual(chump_game.map, None)
+
+    #     # Unfortunately we can't attach sexy beasts to games.
+    #     with self.assertRaises(TypeError) as context:
+    #         chump_game.set_map("Hi I'm a sexy beast.")
+
+    #         self.assertTrue("Unable to attach Map Type obect to Game." in context.exception)
+
+    # """
+    # add_entity_to_map tests.
+    # """
+    # def test_add_entity_to_map_no_map_case(self):
+    #     '''
+    #     This should fail as we can't add an entity to a map
+    #     that doesn't exist.
+    #     '''
+    #     chump_game = Game()
+    #     # Important to note that anyone can be a sexy beast, even the plebians.
+    #     sexy_beast = Plebian()
+
+    #     with self.assertRaises(ValueError) as context:
+    #         chump_game.add_entity_to_map(sexy_beast)
+
+    #         self.assertTrue("There is no map associated with the game!" in context.exception)
+
+    # def test_add_entity_to_map_non_entity_case(self):
+    #     '''
+    #     This should fail as we can't add something to the map that does not
+    #     have an x and y cordinate.
+    #     '''
+    #     chump_game = Game()
+    #     some_plain = Plain()
+    #     some_plain.initialize(10, 10)
+    #     chump_game.set_map(some_plain)
+
+    #     with self.assertRaises(AttributeError) as context:
+    #         chump_game.add_entity_to_map(42 + 300)
+
+    #         self.assertTrue("'int' object has no attribute 'x'" in context.exception)
+
+    # def test_add_entity_to_map_SUCCESS_case(self):
+    #     '''
+    #     Default success case. We should be able to add a character to a map, no problems.
+    #     '''
+    #     chump_game = Game()
+    #     some_plain = Plain()
+    #     some_plain.initialize(10, 10)
+    #     chump_game.set_map(some_plain)
+
+    #     x_cord = 2
+    #     y_cord = 2
+
+    #     the_victim = Plebian()
+    #     chump_game.set_entity_position(the_victim, x_cord, y_cord)
+    #     chump_game.add_entity_to_map(the_victim)
+
+    #     self.assertEqual(
+    #         chump_game.map.grid[x_cord][y_cord].occupant,
+    #         the_victim
+    #     )
+
+    # def test_add_entity_to_map_DUPLICATE_case(self):
+    #     '''
+    #     We're going to add the same character TWICE to the same map.
+    #     Once with the same coordinates, and then once with different coordinates.
+
+    #     The first case should fail (due to two entities being on the same space),
+    #     but the second case should pass (due to them being on different spaces).
+
+    #     I'm okay (for now) with the second case passing because maybe there could
+    #     be "clones" on the map, but this is worth investigation. And I dunno, it
+    #     could get messy.
+    #     '''
+    #     chump_game = Game()
+    #     some_plain = Plain()
+    #     some_plain.initialize(10, 10)
+    #     chump_game.set_map(some_plain)
+
+    #     x_cord = 2
+    #     y_cord = 2
+
+    #     the_victim = Plebian()
+    #     chump_game.set_entity_position(the_victim, x_cord, y_cord)
+    #     chump_game.add_entity_to_map(the_victim)
+
+    #     self.assertEqual(
+    #         chump_game.map.grid[x_cord][y_cord].occupant,
+    #         the_victim
+    #     )
+
+    #     with self.assertRaises(AttributeError) as context:
+    #         chump_game.add_entity_to_map(the_victim)
+
+    #         self.assertTrue("Map space is already occupied!" in context.exception)
+
+    #     # Update the position so we can re-add "the victim" to the map.
+    #     chump_game.set_entity_position(the_victim, x_cord + 1, y_cord + 1)
+    #     chump_game.add_entity_to_map(the_victim)
+
+    #     self.assertEqual(
+    #         chump_game.map.grid[x_cord + 1][y_cord + 1].occupant,
+    #         the_victim
+    #     )
+    #     # Check again to make sure the first "victim" is still on the map.
+    #     self.assertEqual(
+    #         chump_game.map.grid[x_cord][y_cord].occupant,
+    #         the_victim
+    #     )
+
+    # def test_add_entity_to_map_OUT_OF_BOUNDS_case(self):
+    #     '''
+    #     Try to add an entity to a map, but the entity coordinates will
+    #     be out of bounds.
+
+    #     We'll get a key error, but maybe there'd be a way to have more
+    #     rigorous error checking. Maybe have some checks on the Character class?
+    #     Not sure.
+    #     '''
+    #     chump_game = Game()
+    #     some_plain = Plain()
+    #     some_plain.initialize(10, 10)
+    #     chump_game.set_map(some_plain)
+
+    #     x_cord = 12
+    #     y_cord = 12
+
+    #     the_lost_one = Plebian()
+    #     chump_game.set_entity_position(the_lost_one, x_cord, y_cord)
+    #     with self.assertRaises(KeyError):
+    #         chump_game.add_entity_to_map(the_lost_one)
+
+    # """
+    # remove_entity_from_map tests.
+    # """
+    # def test_remove_entity_from_map_SUCCESS_case(self):
+    #     '''
+    #     Put an entity on a map and then remove it. This should be pretty
+    #     straightforward.
+    #     '''
+    #     chump_game = Game()
+    #     some_plain = Plain()
+    #     some_plain.initialize(10, 10)
+    #     chump_game.set_map(some_plain)
+
+    #     x_cord = 2
+    #     y_cord = 2
+
+    #     the_victim = Plebian()
+    #     chump_game.set_entity_position(the_victim, x_cord, y_cord)
+    #     chump_game.add_entity_to_map(the_victim)
+
+    #     # Now we remove the victim. I feel like I'm some mafia member cleaning the scene.
+    #     chump_game.remove_entity_from_map(the_victim)
+    #     self.assertEqual(
+    #         chump_game.map.grid[x_cord][y_cord].occupant,
+    #         None
+    #     )
+
+    # @patch('builtins.print')
+    # def test_remove_entity_from_map_EMPTY_case(self, mock_print):
+    #     '''
+    #     Create an entity, but don't add it to the map. We'll try to remove it, but won't
+    #     be able to as, well, it just doesn't exist.
+    #     '''
+    #     chump_game = Game()
+    #     some_plain = Plain()
+    #     some_plain.initialize(10, 10)
+    #     chump_game.set_map(some_plain)
+
+    #     x_cord = 2
+    #     y_cord = 2
+
+    #     mr_waffles = Plebian()
+    #     chump_game.set_entity_position(mr_waffles, x_cord, y_cord)
+
+    #     chump_game.remove_entity_from_map(mr_waffles)
+    #     mock_print.assert_called_with("{}, {} is already empty!".format(x_cord, y_cord))
+
+    # def test_remove_entity_from_map_NO_MAP_case(self):
+    #     '''
+    #     This should fail, as how can we remove something from a map if a map doesn't exist?
+
+    #     At this point though I feel like I'm testing the compiler more than anything else.
+    #     '''
+    #     chump_game = Game()
+
+    #     x_cord = 2
+    #     y_cord = 2
+
+    #     some_guy = Plebian()
+    #     chump_game.set_entity_position(some_guy, x_cord, y_cord)
+
+    #     with self.assertRaises(ValueError) as context:
+    #         chump_game.remove_entity_from_map(some_guy)
+
+    #         self.assertTrue("There is no map associated with the game!" in context.exception)
+
+    # def test_remove_entity_from_map_non_entity_case(self):
+    #     '''
+    #     This will obviously error. I can't think of many cases where this should naturally
+    #     occur though.
+    #     '''
+    #     chump_game = Game()
+    #     some_plain = Plain()
+    #     some_plain.initialize(10, 10)
+    #     chump_game.set_map(some_plain)
+
+    #     with self.assertRaises(AttributeError) as context:
+    #         chump_game.remove_entity_from_map(666)
+
+    #         self.assertTrue("'int' object has no attribute 'x'" in context.exception)
+
+    # """
+    # check_if_occupied tests.
+    # """
+    # def test_check_if_occupied_NO_MAP_case(self):
+    #     '''
+    #     Hard to check if a tile on a map is occupied if there's no map.
+
+    #     I'm sure someone has said this before. Maybe the voices in my head.
+    #     '''
+    #     chump_game = Game()
+    #     x_cord = 2
+    #     y_cord = 2
+    #     with self.assertRaises(ValueError) as context:
+    #         chump_game.check_if_occupied(x_cord, y_cord)
+
+    #         self.assertTrue("There is no map associated with the game!" in context.exception)
+
+    # def test_check_if_occupied_OUT_OF_BOUNDS_case(self):
+    #     """
+    #     Exactly what is described by the title. We should get a key error.
+    #     """
+    #     the_game = Game()
+    #     smallest_plain = Plain()
+    #     smallest_plain.initialize(3, 3)
+    #     the_game.set_map(smallest_plain)
+
+    #     x_cord = 40000
+    #     y_cord = 40000
+
+    #     with self.assertRaises(KeyError):
+    #         the_game.check_if_occupied(x_cord, y_cord)
+
+    # def test_check_if_occupied_FALSE_case(self):
+    #     """
+    #     Check if an unoccupied square is unoccupied.
+
+    #     Remember, "False" isn't necessarily a bad thing.
+
+    #     "Did you murder this man?"
+
+    #     "False."
+    #     """
+    #     chump_game = Game()
+    #     plane = Plain()
+    #     plane.initialize(10, 10)
+    #     chump_game.set_map(plane)
+
+    #     x_cord = 8
+    #     y_cord = 8
+
+    #     self.assertEqual(
+    #         chump_game.check_if_occupied(x_cord, y_cord),
+    #         False
+    #     )
+
+    # def test_check_if_occupied_TRUE_case(self):
+    #     """
+    #     Check if an occupied square is occupied.
+
+    #     Note that "True" CAN be a bad thing.
+
+    #     "Did you set your boss's house on fire?"
+
+    #     "True."
+    #     """
+    #     chump_game = Game()
+    #     the_void = Plain()
+    #     the_void.initialize(10, 10)
+    #     chump_game.set_map(the_void)
+
+    #     x_cord = 3
+    #     y_cord = 3
+
+    #     master_of_the_universe = Plebian()
+    #     chump_game.set_entity_position(master_of_the_universe, x_cord, y_cord)
+    #     chump_game.add_entity_to_map(master_of_the_universe)
+
+    #     self.assertEqual(
+    #         chump_game.check_if_occupied(x_cord, y_cord),
+    #         True
+    #     )
+
+    # """
+    # initialize_map tests.
+    # """
+    # @patch('classes.game.input', create=True)
+    # def test_initialize_map_SUCCESS_case(self, mocked_input):
+    #     """
+    #     Simple case of initializing a map and assigning it to the game.
+
+    #     Note we're going to have to supply user inputs for this.
+
+    #     Furthermore, this is just to make sure the map gets attached to the game,
+    #     we can be more rigorous with the "initialize" method on the Maps.
+    #     """
+    #     x_cord = 10
+    #     y_cord = 8
+    #     mocked_input.side_effect = [x_cord, y_cord]
+    #     a_game = Game()
+    #     a_game.initialize_map()
+
+    #     self.assertTrue(isinstance(a_game.map, Plain))
+    #     self.assertEqual(a_game.map.width, x_cord)
+    #     self.assertEqual(a_game.map.height, y_cord)
+
+    # @patch('classes.game.input', create=True)
+    # def test_initialize_map_FAILURE_case(self, mocked_input):
+    #     """
+    #     Going to crash this method by giving it some string inputs.
+
+    #     This raises the question of maybe we should allow the user to try
+    #     again if they input something "dumb", but that might be a later task.
+    #     """
+    #     x_cord = "ten"
+    #     y_cord = "eight"
+    #     mocked_input.side_effect = [x_cord, y_cord]
+    #     a_game = Game()
+
+    #     with self.assertRaises(ValueError):
+    #         a_game.initialize_map()
 
     """
     create_character tests.
@@ -550,7 +569,9 @@ class TestGame(unittest.TestCase):
     """
     move_character tests.
     """
-    @patch('classes.game.input', create=True)
+    # @patch('classes.game.input', create=True)
+    # @patch('classes.mapengine.input', create=True) # mocked_input_mapengine, mocked_input_game
+    @patch('builtins.input', create=True)
     def test_move_character_success_case(self, mocked_input):
         """
         Simple test to see if we can move a character on the map. Note that the movement
@@ -568,6 +589,10 @@ class TestGame(unittest.TestCase):
         down = "down"
         left = "left"
         right = "right"
+        # mocked_input_mapengine.side_effect = [
+
+
+        # ]
         mocked_input.side_effect = [
             map_width,
             map_height,
@@ -588,46 +613,50 @@ class TestGame(unittest.TestCase):
 
         # Set up to put the character on the map
         chump_game = Game()
-        chump_game.initialize_map()
+        the_mapengine = MapEngine()
+        chump_game.set_map_engine(the_mapengine)
+        the_mapengine.initialize_map()
         plebian_isaac = chump_game.create_character()
         chump_game.set_character_position(plebian_isaac)
-        chump_game.add_entity_to_map(plebian_isaac)
+        the_mapengine.add_entity_to_map(plebian_isaac)
 
         # Make sure Isaac is on the map.
         self.assertEqual(
-            chump_game.map.grid[int(char_starting_x)][int(char_starting_y)].occupant,
+            the_mapengine.map.grid[int(char_starting_x)][int(char_starting_y)].occupant,
             plebian_isaac
         )
 
         # Move Isaac "Up"
         chump_game.move_character()
         self.assertEqual(
-            chump_game.map.grid[int(char_starting_x)][int(char_starting_y) + 1].occupant,
+            the_mapengine.map.grid[int(char_starting_x)][int(char_starting_y) + 1].occupant,
             plebian_isaac
         )
 
         # Move Isaac "Down"
         chump_game.move_character()
         self.assertEqual(
-            chump_game.map.grid[int(char_starting_x)][int(char_starting_y)].occupant,
+            the_mapengine.map.grid[int(char_starting_x)][int(char_starting_y)].occupant,
             plebian_isaac
         )
 
         # Move Isaac "Left"
         chump_game.move_character()
         self.assertEqual(
-            chump_game.map.grid[int(char_starting_x) - 1][int(char_starting_y)].occupant,
+            the_mapengine.map.grid[int(char_starting_x) - 1][int(char_starting_y)].occupant,
             plebian_isaac
         )
 
         # Move Isaac "Right"
         chump_game.move_character()
         self.assertEqual(
-            chump_game.map.grid[int(char_starting_x)][int(char_starting_y)].occupant,
+            the_mapengine.map.grid[int(char_starting_x)][int(char_starting_y)].occupant,
             plebian_isaac
         )
 
-    @patch('classes.game.input', create=True)
+    # @patch('classes.game.input', create=True) # mocked_input_mapengine, mocked_input_game
+    # @patch('classes.mapengine.input', create=True)
+    @patch('builtins.input', create=True)
     def test_move_character_non_existent_character_case(self, mocked_input):
         """
         Try to move a character who just doesn't exist. Big sad.
@@ -641,16 +670,23 @@ class TestGame(unittest.TestCase):
             map_height,
             character_name,
         ]
+        # mocked_input_game.side_effect = [
+        #     character_name,
+        # ]
 
         chump_game = Game()
-        chump_game.initialize_map()
+        the_mapengine = MapEngine()
+        chump_game.set_map_engine(the_mapengine)
+        the_mapengine.initialize_map()
 
         with self.assertRaises(ValueError) as context:
             chump_game.move_character()
 
             self.assertTrue("{} is not a character that exists on the map!".format(character_name) in context.exception)
 
-    @patch('classes.game.input', create=True)
+    # @patch('classes.game.input', create=True) # mocked_input_mapengine, mocked_input_game
+    # @patch('classes.mapengine.input', create=True)
+    @patch('builtins.input', create=True)
     def test_move_character_already_occupied_case(self, mocked_input):
         """
         Now we try to move a chracter to a spot that is already occupied.
@@ -684,26 +720,31 @@ class TestGame(unittest.TestCase):
             character_1_name,
             up,
         ]
+        # mocked_input_game.side_effect = [
+
+        # ]
 
         # Set up to put the character on the map
         chump_game = Game()
-        chump_game.initialize_map()
+        the_mapengine = MapEngine()
+        chump_game.set_map_engine(the_mapengine)
+        the_mapengine.initialize_map()
 
         # Add Isaac to the map and make sure he's there.
         plebian_isaac = chump_game.create_character()
         chump_game.set_character_position(plebian_isaac)
-        chump_game.add_entity_to_map(plebian_isaac)
+        the_mapengine.add_entity_to_map(plebian_isaac)
         self.assertEqual(
-            chump_game.map.grid[int(char_1_starting_x)][int(char_1_starting_y)].occupant,
+            the_mapengine.map.grid[int(char_1_starting_x)][int(char_1_starting_y)].occupant,
             plebian_isaac
         )
 
         # Add Darius to the map and make sure he's there.
         jester_darius = chump_game.create_character()
         chump_game.set_character_position(jester_darius)
-        chump_game.add_entity_to_map(jester_darius)
+        the_mapengine.add_entity_to_map(jester_darius)
         self.assertEqual(
-            chump_game.map.grid[int(char_2_starting_x)][int(char_2_starting_y)].occupant,
+            the_mapengine.map.grid[int(char_2_starting_x)][int(char_2_starting_y)].occupant,
             jester_darius
         )
 
@@ -711,15 +752,17 @@ class TestGame(unittest.TestCase):
         # As such Isaac should still be in the same spot. Same as Darius, for that matter.
         chump_game.move_character()
         self.assertEqual(
-            chump_game.map.grid[int(char_1_starting_x)][int(char_1_starting_y)].occupant,
+            the_mapengine.map.grid[int(char_1_starting_x)][int(char_1_starting_y)].occupant,
             plebian_isaac
         )
         self.assertEqual(
-            chump_game.map.grid[int(char_2_starting_x)][int(char_2_starting_y)].occupant,
+            the_mapengine.map.grid[int(char_2_starting_x)][int(char_2_starting_y)].occupant,
             jester_darius
         )
 
-    @patch('classes.game.input', create=True)
+    # @patch('classes.game.input', create=True) # mocked_input_mapengine, mocked_input_game
+    # @patch('classes.mapengine.input', create=True)
+    @patch('builtins.input', create=True)
     def test_move_character_NO_MAP_case(self, mocked_input):
         """
         Try to move a character when we have no map attached to the game. As expected this
@@ -740,9 +783,14 @@ class TestGame(unittest.TestCase):
             character_name,
             up,
         ]
+        # mocked_input_game.side_effect = [
+
+        # ]
 
         # Set up to put the character on the map
         chump_game = Game()
+        the_mapengine = MapEngine()
+        chump_game.set_map_engine(the_mapengine)
         plebian_isaac = chump_game.create_character()
         chump_game.set_character_position(plebian_isaac)
         # Normally the error would occur here, but that's not what we're trying to test.
@@ -753,7 +801,9 @@ class TestGame(unittest.TestCase):
 
             self.assertTrue("There is no map associated with the game!" in context.exception)
 
-    @patch('classes.game.input', create=True)
+    # @patch('classes.game.input', create=True) # mocked_input_mapengine, mocked_input_game
+    # @patch('classes.mapengine.input', create=True)
+    @patch('builtins.input', create=True)
     def test_move_character_OUT_OF_BOUNDS_case(self, mocked_input):
         """
         Try to move a character out of bounds - a message should be printed (though we won't see it),
@@ -778,17 +828,23 @@ class TestGame(unittest.TestCase):
             character_name,
             down,
         ]
+        # mocked_input_game.side_effect = [
+
+
+        # ]
 
         # Set up to put the character on the map
         chump_game = Game()
-        chump_game.initialize_map()
+        the_mapengine = MapEngine()
+        chump_game.set_map_engine(the_mapengine)
+        the_mapengine.initialize_map()
         plebian_isaac = chump_game.create_character()
         chump_game.set_character_position(plebian_isaac)
-        chump_game.add_entity_to_map(plebian_isaac)
+        the_mapengine.add_entity_to_map(plebian_isaac)
 
         # Make sure Isaac is on the map.
         self.assertEqual(
-            chump_game.map.grid[int(char_starting_x)][int(char_starting_y)].occupant,
+            the_mapengine.map.grid[int(char_starting_x)][int(char_starting_y)].occupant,
             plebian_isaac
         )
 
@@ -796,66 +852,66 @@ class TestGame(unittest.TestCase):
         # should stay the same.
         chump_game.move_character()
         self.assertEqual(
-            chump_game.map.grid[int(char_starting_x)][int(char_starting_y)].occupant,
+            the_mapengine.map.grid[int(char_starting_x)][int(char_starting_y)].occupant,
             plebian_isaac
         )
 
-    """
-    set_maptile_occupant tests
-    """
-    def test_set_maptile_occupant_success_case(self):
-        """
-        Basic success case. If this fails we're in a heap of trouble.
-        """
-        heroic_maptile = MapTile()
-        simply_dan = Plebian()
-        chump_game = Game()
+    # """
+    # set_maptile_occupant tests
+    # """
+    # def test_set_maptile_occupant_success_case(self):
+    #     """
+    #     Basic success case. If this fails we're in a heap of trouble.
+    #     """
+    #     heroic_maptile = MapTile()
+    #     simply_dan = Plebian()
+    #     chump_game = Game()
 
-        # No occupant... YET
-        self.assertEqual(heroic_maptile.occupant, None)
+    #     # No occupant... YET
+    #     self.assertEqual(heroic_maptile.occupant, None)
 
-        chump_game.set_maptile_occupant(simply_dan, heroic_maptile)
+    #     chump_game.set_maptile_occupant(simply_dan, heroic_maptile)
 
-        # Now Dan is an occupant.
-        self.assertEqual(heroic_maptile.occupant, simply_dan)
+    #     # Now Dan is an occupant.
+    #     self.assertEqual(heroic_maptile.occupant, simply_dan)
 
-    def test_set_maptile_occupant_n0ne_case(self):
-        """
-        In my infinite wisdom I forgot that we want to be able to set
-        the map_occupant to be "None" - a relevant case for when we
-        remove enities from map.
-        """
-        heroic_maptile = MapTile()
-        simply_dan = Plebian()
-        chump_game = Game()
+    # def test_set_maptile_occupant_n0ne_case(self):
+    #     """
+    #     In my infinite wisdom I forgot that we want to be able to set
+    #     the map_occupant to be "None" - a relevant case for when we
+    #     remove enities from map.
+    #     """
+    #     heroic_maptile = MapTile()
+    #     simply_dan = Plebian()
+    #     chump_game = Game()
 
-        # No occupant... YET
-        self.assertEqual(heroic_maptile.occupant, None)
+    #     # No occupant... YET
+    #     self.assertEqual(heroic_maptile.occupant, None)
 
-        chump_game.set_maptile_occupant(simply_dan, heroic_maptile)
+    #     chump_game.set_maptile_occupant(simply_dan, heroic_maptile)
 
-        # Dan is an occupant. All is good.
-        self.assertEqual(heroic_maptile.occupant, simply_dan)
+    #     # Dan is an occupant. All is good.
+    #     self.assertEqual(heroic_maptile.occupant, simply_dan)
 
-        # Now we don't want the occupant to be anything.
-        chump_game.set_maptile_occupant(None, heroic_maptile)
+    #     # Now we don't want the occupant to be anything.
+    #     chump_game.set_maptile_occupant(None, heroic_maptile)
 
-        # There's nothing here now!
-        self.assertEqual(heroic_maptile.occupant, None)
+    #     # There's nothing here now!
+    #     self.assertEqual(heroic_maptile.occupant, None)
 
-    def test_set_maptile_occupant_FAILURE_case(self):
-        """
-        We're going to try setting an occupant that isn't a character or is "None".
-        This shouldn't fly, though we may have to make adjustments later.
-        """
-        heroic_maptile = MapTile()
-        epic_string = "I am the ember of eternity!"
-        chump_game = Game()
+    # def test_set_maptile_occupant_FAILURE_case(self):
+    #     """
+    #     We're going to try setting an occupant that isn't a character or is "None".
+    #     This shouldn't fly, though we may have to make adjustments later.
+    #     """
+    #     heroic_maptile = MapTile()
+    #     epic_string = "I am the ember of eternity!"
+    #     chump_game = Game()
 
-        with self.assertRaises(TypeError) as context:
-            chump_game.set_maptile_occupant(epic_string, heroic_maptile)
+    #     with self.assertRaises(TypeError) as context:
+    #         chump_game.set_maptile_occupant(epic_string, heroic_maptile)
 
-            self.assertTrue("Cannot make a non-character a map occupant!" in context.exception)
+    #         self.assertTrue("Cannot make a non-character a map occupant!" in context.exception)
 
     """
     set_entity_position tests
